+++ CHANGELOG.md
@@ -1,8 +1,19 @@
 # Just Hammers Changelog
 
-## [2.0.1]
+## [2.0.2]
+
+### Added
+
+- A nice new tooltip to show the durability of the hammer.
+- You can now shift to negate the hammer's ability to break blocks in the AOE (Area of Effect).
+
+### Changed
+
+- All hammers above the basic 3x3x1 (Not requiring netherite) now have at least the durability of the netherite pickaxe.
+  - This is a pretty major bump in durability for the hammers, but it's a lot more fun to use them now.
 
 ### Fixed
 
-- Forge compatibility with modded enchantments. Thanks to ([#11](https://github.com/ErrorMikey/JustHammers/pull/11)) [[allanf181](https://github.com/allanf181)]
-- Hammers able to break blocks higher than their tier and voiding those blocks. Thanks to ([#10](https://github.com/ErrorMikey/JustHammers/pull/10)) [[Sisko](https://github.com/CaptainSisko)]
+- Hammers no longer break! This was intended from the start but this actually works now. If your hammer is about to break, it will stop being able to mine until it's repaired using an anvil.
+- Hammers no longer do weird things when looking at specific angles. Thanks to [#14](https://github.com/ErrorMikey/JustHammers/issues/14) [@danielhoyosceldran](https://github.com/danielhoyosceldran)
+- Hammers work again in creative mode when on Fabric [#16](https://github.com/ErrorMikey/JustHammers/issues/16) [@prshs70](https://github.com/prshs70)